requirements.txt
openai
python-dotenv
groq
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-groq